# api/internal/handlers/swagger.yaml
openapi: 3.0.3
info:
  title: StreamSpace API
  description: |
    StreamSpace is a platform-agnostic container streaming platform that delivers GUI applications to web browsers.

    ## Architecture
    - **Control Plane**: Centralized API with WebSocket Hub for agent coordination
    - **Agents**: Lightweight executors on target platforms (Kubernetes, Docker)
    - **VNC Proxy**: Secure, firewall-friendly VNC streaming through WebSocket

    ## Authentication
    - **Users**: JWT tokens (24-hour expiration) via `/api/v1/auth/login`
    - **Agents**: API keys via `/api/v1/agents/register`
    - **Webhooks**: HMAC signatures

    ## Rate Limiting
    Default: 60 requests/minute per IP (configurable via `RATE_LIMIT_REQUESTS_PER_MINUTE`)

    ## Common Response Formats
    All error responses follow:
    ```json
    {
      "error": "ErrorCode",
      "message": "Detailed error message"
    }
    ```
  version: 2.0.0-beta
  contact:
    name: StreamSpace Team
    url: https://github.com/streamspace-dev/streamspace
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.streamspace.example.com
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for agent authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
      required:
        - error
        - message
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user:
          type: string
          description: Username owning the session
        template:
          type: string
          description: Template name
        status:
          type: string
          enum: [pending, creating, running, hibernating, hibernated, stopping, stopped, failed, terminated]
        agent_id:
          type: string
          description: ID of the agent running this session
        vnc_port:
          type: integer
          description: VNC port (internal to agent)
        resources:
          $ref: '#/components/schemas/Resources'
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateSessionRequest:
      type: object
      required:
        - user
      properties:
        user:
          type: string
          description: Username for the session
        template:
          type: string
          description: Template name to use
        application_id:
          type: string
          description: Installed application ID (alternative to template)
        resources:
          $ref: '#/components/schemas/Resources'
        persistent_home:
          type: boolean
          default: true
          description: Mount persistent home directory
        idle_timeout:
          type: string
          example: "30m"
          description: Auto-hibernate after idle period
        max_session_duration:
          type: string
          example: "8h"
          description: Maximum session lifetime
        tags:
          type: array
          items:
            type: string
    Resources:
      type: object
      properties:
        memory:
          type: string
          example: "2Gi"
        cpu:
          type: string
          example: "1000m"
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer
paths:
...
  /api/v1/sessions:
    get:
      summary: List sessions
      description: Get all sessions with optional filtering
      operationId: listSessions
      tags:
        - sessions
      security:
        - bearerAuth: []
      parameters:
        - name: user
          in: query
          schema:
            type: string
          description: Filter by username
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, creating, running, hibernating, hibernated, stopping, stopped, failed, terminated]
        - name: template
          in: query
          schema:
            type: string
        - name: agent_id
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Session'

    post:
      summary: Create session
      description: Create a new container session
      operationId: createSession
      tags:
        - sessions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Quota exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/sessions/{id}:
    get:
      summary: Get session
      operationId: getSession
      tags:
        - sessions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '404':
          description: Session not found

    patch:
      summary: Update session
      operationId: updateSession
      tags:
        - sessions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                resources:
                  $ref: '#/components/schemas/Resources'
                idle_timeout:
                  type: string
      responses:
        '200':
          description: Session updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'

    delete:
      summary: Delete session
      description: Terminate and delete a session
      operationId: deleteSession
      tags:
        - sessions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Session deleted
        '404':
          description: Session not found
...
  /api/v1/users/{id}/sessions:
    get:
      summary: Get user's sessions
      operationId: getUserSessions
      tags:
        - users
        - sessions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User's sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'

  # Groups
...